+++ views/partials/preview.ejs
@@ -28,3 +28,4 @@
     </div>
 
 </div>
+<script src="/js/preview.js" type="module"></script>
